+++ src/rpc/blockchain.cpp
@@ -4,8 +4,8 @@
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
 #include "main.h"
-#include "rpcserver.h"
-#include "rpcprotocol.h"
+#include "server.h"
+#include "protocol.h"
 #include "init.h" // for pwalletMain
 #include "checkpoints.h"
 #include "txdb.h"
