+++ src/rpc/net.cpp
@@ -2,8 +2,8 @@
 // Distributed under the MIT/X11 software license, see the accompanying
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
-#include "rpcserver.h"
-#include "rpcprotocol.h"
+#include "server.h"
+#include "protocol.h"
 #include "alert.h"
 #include "wallet/wallet.h"
 #include "wallet/db.h"
