+++ src/rpc/rawtransaction.cpp
@@ -14,8 +14,8 @@
 #include "gridcoin/support/block_finder.h"
 #include "gridcoin/tx_message.h"
 #include "gridcoin/voting/payloads.h"
-#include "rpcprotocol.h"
-#include "rpcserver.h"
+#include "protocol.h"
+#include "server.h"
 #include "streams.h"
 #include "txdb.h"
 #include "wallet/coincontrol.h"
