+++ src/wallet/wallet.cpp
@@ -12,9 +12,9 @@
 #include "wallet/coincontrol.h"
 #include <boost/algorithm/string/replace.hpp>
 #include <boost/thread.hpp>
-#include "rpcserver.h"
-#include "rpcclient.h"
-#include "rpcprotocol.h"
+#include "rpc/server.h"
+#include "rpc/client.h"
+#include "rpc/protocol.h"
 #include <boost/variant/apply_visitor.hpp>
 #include <script.h>
 #include "main.h"
